added check for tmp dir null before deleting
git-svn-id: https://ncisvn.nci.nih.gov/svn/cagrid/trunk@2966 dd7b0c16-3c94-4492-92de-414b6a06f0b1

+++ cagrid-1-0/tests/projects/coretests/src/java/gov/nci/nih/cagrid/tests/core/GlobusHelper.java
@@ -206,7 +206,7 @@ public class GlobusHelper
 	
 	public void cleanupTempGlobus()
 	{
-		FileUtils.deleteRecursive(tmpGlobusLocation);
+		if (tmpGlobusLocation != null) FileUtils.deleteRecursive(tmpGlobusLocation);
 	}
 	
 	private static void sleep(long ms)
